+++ sentry/src/main/java/io/sentry/Sentry.java
@@ -11,6 +11,7 @@ import io.sentry.internal.modules.IModulesLoader;
 import io.sentry.internal.modules.ManifestModulesLoader;
 import io.sentry.internal.modules.NoOpModulesLoader;
 import io.sentry.internal.modules.ResourcesModulesLoader;
+import io.sentry.metrics.MetricsApi;
 import io.sentry.protocol.SentryId;
 import io.sentry.protocol.User;
 import io.sentry.transport.NoOpEnvelopeCache;
@@ -973,6 +974,13 @@ public final class Sentry {
   @SuppressWarnings("InlineMeSuggester")
   public static void reportFullDisplayed() {
     reportFullyDisplayed();
+  }
+
+  /** the metrics API for the current hub */
+  @NotNull
+  @ApiStatus.Experimental
+  public static MetricsApi metrics() {
+    return getCurrentHub().metrics();
   }
 
   /**